pyproject.toml
[project]
name = "flighttracker"
version = "0.1.0"
description = "Weekly flight digest for a FlightRadar24 feeder station"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9",
    "asyncpg>=0.29",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
